[#55] refactor : atomKeys.ts 변경
