+++ src/components/BottomSheetModal/ItemBrandSelectModalComponent/HotBrand.tsx
@@ -6,8 +6,10 @@ import useModals from '../../Modals/hooks/useModals'
 import { modals } from '../../Modals'
 import Chip from '../../Chip/Chip'
 import { ChipWrapper } from '../ItemBrandSelectModal'
-import { Brand } from '../../../pages/item/create/components/BrandItemField/BrandItemField'
-import { selectedBrandState } from '../../../config/atomKeys'
+import {
+  Brand,
+  selectedBrandState,
+} from '../../../pages/item/create/components/BrandItemField/BrandItemField'
 
 const HotBrand = () => {
   const brandList = [
